shift preferences utils under ts/utils/profile
